feat: 신규 클래스 생성
